+++ src/components/Home/Authenticated/AuthenticatedContainer.tsx
@@ -1,0 +1,5 @@
+import { TripsContainer } from "./TripsContainer";
+
+export const Authenticatedcontainer = () => {
+  return <TripsContainer />;
+};
