+++ src/components/Home/Authenticated/TripsContainer.tsx
