+++ src/components/Home/HomeContainer.tsx
@@ -1,9 +1,6 @@
+import { UnauthenticatedContainer } from "./Unauthenticated/UnauthenticatedContainer";
+import { Authenticatedcontainer } from "./Authenticated/AuthenticatedContainer";
 import { useSession } from "next-auth/react";
-import { GetStartedButton } from "./GetStartedButton";
-import { HeroBanner } from "./HeroBanner";
-import clsx from "clsx";
-import { TripsContainer } from "./TripsContainer";
-import { Footer } from "../Footer";
 
 export const HomeContainer = () => {
   const { status } = useSession();
@@ -11,29 +8,9 @@ export const HomeContainer = () => {
   return (
     <>
       {status === "authenticated" ? (
-        <div className={clsx("p-5")}>
-          <TripsContainer />
-        </div>
+        <Authenticatedcontainer />
       ) : (
-        <div
-          className={clsx(
-            "flex flex-col",
-            "h-[calc(100vh-70px)] w-screen",
-            "items-center justify-center",
-          )}
-        >
-          <div className={clsx("p-5")}>
-            <HeroBanner />
-          </div>
-
-          <div className={clsx("p-5")}>
-            <GetStartedButton />
-          </div>
-
-          <div className={clsx("absolute bottom-0 lg:bottom-3")}>
-            <Footer />
-          </div>
-        </div>
+        <UnauthenticatedContainer />
       )}
     </>
   );
